arpの下降時に音が鳴らない不具合修正

+++ src/cmd/txt2seq_ana.rs
@@ -122,7 +122,7 @@ fn arp_translation(mut beat_analysis: Vec<Vec<i16>>, exps: &Vec<String>) -> Vec<
           (last_note as i32)-(crnt_note as i32) < 10 &&
           (crnt_note as i32)-(last_note as i32) < 10 {
             // 過去＆現在：単音、ノート適正、差が10半音以内
-            ana.push((crnt_note-last_note) as i16); // arp
+            ana.push(crnt_note as i16-last_note as i16); // arp
         }
         else {
             ana.push(ARP_COM);    // com
